fix: now rescaling only img and now it is module
